requirements.txt
asgiref==3.7.2
Django==5.0.2
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
memory-profiler==0.61.0
psutil==5.9.8
psycopg==3.1.18
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
